+++ lib/rubocop/cop/bundler/ordered_gems.rb
@@ -54,7 +54,7 @@ module RuboCop
         end
 
         def_node_search :gem_declarations, <<-PATTERN
-          (:send nil? :gem ...)
+          (:send nil? :gem str ...)
         PATTERN
       end
     end