+++ spec/rubocop/cop/bundler/ordered_gems_spec.rb
@@ -24,6 +24,24 @@ describe RuboCop::Cop::Bundler::OrderedGems, :config do
     end
   end
 
+  context 'when a gem is referenced from a variable' do
+    it 'ignores the line' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        gem 'rspec'
+        gem ENV['env_key_undefined'] if ENV.key?('env_key_undefined')
+        gem 'rubocop'
+      RUBY
+    end
+
+    it 'resets the sorting to a new block' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        gem 'rubocop'
+        gem ENV['env_key_undefined'] if ENV.key?('env_key_undefined')
+        gem 'ast'
+      RUBY
+    end
+  end
+
   context 'When gems are not alphabetically sorted' do
     let(:source) { <<-RUBY.strip_indent }
       gem 'rubocop'